requirements.txt
PyGithub==2.1.1
GitPython==3.1.40
PyYAML==6.0.1
requests==2.31.0

